8/pyproject.toml
[tool.poetry]
name = "exercise"
version = "0.1.0"
description = "exercise"
authors = ["exercise"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.0.1"
asyncpg = "^0.30.0"
asyncio = "^3.4.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
greenlet = "^3.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
